requirements.txt
rich
mysql-connector-python==8.0.29